+++ cli/CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial CLI is built and usable in it's most basic form. Listed endpoints can be scrolled.
 - Transaction payments can be seen
+- Transaction details can be seen